Organize arr service modules into dedicated subpackages
